+++ src/appInterface.h
@@ -41,6 +41,7 @@ class IApp {
 
         virtual uint32_t getUptime() = 0;
         virtual uint32_t getTimestamp() = 0;
+        virtual uint64_t getTimestampMs() = 0;
         virtual uint32_t getSunrise() = 0;
         virtual uint32_t getSunset() = 0;
         virtual void setTimestamp(uint32_t newTime) = 0;