+++ src/utils/helper.cpp
@@ -70,12 +70,14 @@ namespace ah {
         return String(str);
     }
 
-    String getTimeStrMs(time_t t) {
+    String getTimeStrMs(uint64_t t) {
         char str[13];
         if(0 == t)
             sprintf(str, "n/a");
-        else
+        else {
+            t = (t + (millis() % 1000)) / 1000;
             sprintf(str, "%02d:%02d:%02d.%03d", hour(t), minute(t), second(t), millis() % 1000);
+        }
         return String(str);
     }